refactor(api): move hook handlers to new package (#829)
